+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/levelgen/feature/configured/BWGOverworldVegetationConfiguredFeatures.java
@@ -449,7 +449,6 @@ public class BWGOverworldVegetationConfiguredFeatures {
     }
 
 
-
     public static void init() {
         BiomesWeveGone.LOGGER.info("Creating and Registering Overworld Vegetation Configured Features");
     }