+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add Missing Spirit Roots LootTable/Drops
 - BoneMealing The Hydrange Hedge block will now Drop Hydrangea Hedges
 - Fix BWG Planks not being shown to craft specific wood set crafting tables
+- BoneMealing Grassblocks in Allium Shrubland, Amaranth Grassland, Coconino Meadow, Crimson Tundra, Enchanted Tangle, Orchard, Pumpkin Valley, Sakura Grove, Temperate Grove will have the chance of flowers from that biome growing
 
 # 2.3.0 -> Major Internal Changes and Refactors
 - Add Spanish (Chile) Translations (es_cl) (Credits: Ganbare-Lucifer)